backend/requirements.txt
fastapi
uvicorn
langchain
pgvector
psycopg2-binary
python-dotenv
pydantic
langchain_community
langchain-text-splitters
langchain-huggingface 
sentence-transformers   
python-multipart
pdfplumber
dashscope
langchain-postgres